+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 - Add the `Password` dialog box.
 - Add the `Zenity` backend.
+- Implement `std::fmt::Display` for the `Error` enum.
 
 # v0.1.0 (2019-01-08)
 - Initial release with the `Input`, `Message` and `Question` dialog boxes and